Reorganise type-erasure code to make parts reusable.
Doing this in preparation form adding more flavours of
type-erasing wrapper.

- Move vtable, vtable_entry and vtable holders into new vtable.hpp header
- Move any_ref to its own header.
- Move with_type_erased_tag_invoke to its own header.
- Change 'swap' CPO to be defined using inline namespace rather
  than an explicit 'using' to get around clang error complaining about
  declaring a friend function with the same name as a 'using'
  declaration.
- Added implicit cast from compatible vtable holders to an
  inline_vtable_holder which now copies the vtable entries.
  This should allow casting a type-erased type to a type-erased
  type with a narrower interface if it uses inline_vtable_holder.
- Removed vtable_holder alias for deciding whether to use
  inline or indirect vtable and just do this inline in each
  type-erasing wrapper class.
