+++ include/unifex/any_unique.hpp
@@ -21,6 +21,8 @@
 #include <unifex/this.hpp>
 #include <unifex/type_traits.hpp>
 #include <unifex/std_concepts.hpp>
+#include <unifex/detail/vtable.hpp>
+#include <unifex/detail/with_type_erased_tag_invoke.hpp>
 
 #include <memory>
 #include <utility>
@@ -29,199 +31,6 @@
 
 namespace unifex {
 namespace _any_unique {
-
-template <typename CPO>
-using base_cpo_t = _overload::base_cpo_t<CPO>;
-
-template <typename CPO, typename T, bool NoExcept, typename Ret, typename... Args>
-Ret _invoke(
-    base_cpo_t<CPO> cpo,
-    replace_this_with_void_ptr_t<Args>... args) noexcept(NoExcept) {
-  static_assert(!NoExcept || noexcept(extract_this<Args...>{}(args...)));
-  void* thisPointer = extract_this<Args...>{}(args...);
-  static_assert(!NoExcept || noexcept(Ret(((base_cpo_t<CPO>&&) cpo)(
-      replace_this<Args>::get(
-          (decltype(args)&&) args,
-          *static_cast<T*>(thisPointer))...))));
-  return ((base_cpo_t<CPO>&&) cpo)(
-      replace_this<Args>::get(
-          (decltype(args)&&) args,
-          *static_cast<T*>(thisPointer))...);
-}
-
-template <typename CPO, typename Sig = typename CPO::type_erased_signature_t>
-struct vtable_entry;
-
-template <typename CPO, typename Ret, typename... Args>
-struct vtable_entry<CPO, Ret(Args...) noexcept> {
-  using fn_t =
-      Ret(base_cpo_t<CPO>, replace_this_with_void_ptr_t<Args>...) noexcept;
-
-  constexpr fn_t* get() const noexcept {
-    return fn_;
-  }
-
-  template <typename T>
-  static constexpr vtable_entry create() noexcept {
-    return vtable_entry{_any_unique::_invoke<CPO, T, true, Ret, Args...>};
-  }
-
-private:
-  explicit constexpr vtable_entry(fn_t* fn) noexcept
-    : fn_(fn) {}
-
-  fn_t* fn_;
-};
-
-template <typename CPO, typename Ret, typename... Args>
-struct vtable_entry<CPO, Ret(Args...)> {
-  using fn_t =
-      Ret(base_cpo_t<CPO>, replace_this_with_void_ptr_t<Args>...);
-
-  constexpr fn_t* get() const noexcept {
-    return fn_;
-  }
-
-  template <typename T>
-  static constexpr vtable_entry create() noexcept {
-    return vtable_entry{_any_unique::_invoke<CPO, T, false, Ret, Args...>};
-  }
-
-private:
-  explicit constexpr vtable_entry(fn_t* fn) noexcept
-    : fn_(fn) {}
-
-  fn_t* fn_;
-};
-
-template <typename... CPOs>
-struct vtable : private vtable_entry<CPOs>... {
-  template <typename T>
-  static constexpr vtable create() noexcept {
-    return vtable{vtable_entry<CPOs>::template create<T>()...};
-  }
-
-  template <typename CPO>
-  constexpr auto get() const noexcept -> typename vtable_entry<CPO>::fn_t* {
-    const vtable_entry<CPO>& entry = *this;
-    return entry.get();
-  }
-
-private:
-  explicit constexpr vtable(vtable_entry<CPOs>... entries) noexcept
-    : vtable_entry<CPOs>{entries}... {}
-};
-
-template <typename... CPOs>
-struct indirect_vtable_holder {
-  template <typename T>
-  static indirect_vtable_holder create() {
-    static constexpr vtable<CPOs...> v = vtable<CPOs...>::template create<T>();
-    return indirect_vtable_holder{v};
-  }
-
-  const vtable<CPOs...>& operator*() const noexcept {
-    return *vtable_;
-  }
-
-  const vtable<CPOs...>* operator->() const noexcept {
-    return vtable_;
-  }
-
- private:
-  constexpr indirect_vtable_holder(const vtable<CPOs...>& vtable)
-    : vtable_(&vtable) {}
-
-  const vtable<CPOs...>* vtable_;
-};
-
-template <typename... CPOs>
-struct inline_vtable_holder {
-  template <typename T>
-  static constexpr inline_vtable_holder create() {
-    return inline_vtable_holder{vtable<CPOs...>::template create<T>()};
-  }
-
-  const vtable<CPOs...>& operator*() const noexcept {
-    return vtable_;
-  }
-
-  const vtable<CPOs...>* operator->() const noexcept {
-    return &vtable_;
-  }
-
- private:
-  constexpr inline_vtable_holder(const vtable<CPOs...>& vtable)
-    : vtable_(vtable) {}
-
-  vtable<CPOs...> vtable_;
-};
-
-template <typename... CPOs>
-using vtable_holder = conditional_t<
-    (sizeof...(CPOs) <= 4),
-    inline_vtable_holder<CPOs...>,
-    indirect_vtable_holder<CPOs...>>;
-
-template <
-    typename Derived,
-    typename CPO,
-    bool NoExcept,
-    typename Sig>
-struct _with_type_erased_tag_invoke;
-
-template <
-    typename Derived,
-    typename CPO,
-    bool NoExcept = false,
-    typename Sig = typename CPO::type_erased_signature_t>
-using with_type_erased_tag_invoke =
-    typename _with_type_erased_tag_invoke<Derived, CPO, NoExcept, Sig>::type;
-
-template <
-    typename Derived,
-    typename CPO,
-    bool NoExcept,
-    typename Ret,
-    typename... Args>
-struct _with_type_erased_tag_invoke<
-    Derived,
-    CPO,
-    NoExcept,
-    Ret(Args...)> {
-  struct type {
-    friend Ret tag_invoke(base_cpo_t<CPO> cpo, replace_this_t<Args, Derived>... args)
-        noexcept(NoExcept) {
-      using cpo_t = base_cpo_t<CPO>;
-      static_assert(
-          !NoExcept || noexcept(extract_this<Args...>{}((decltype(args) &&) args...)));
-      auto&& t = extract_this<Args...>{}((decltype(args) &&) args...);
-      static_assert(!NoExcept || noexcept(get_object_address(t)));
-      void* objPtr = get_object_address(t);
-      static_assert(!NoExcept || noexcept(get_vtable(t)->template get<CPO>()));
-      auto* fnPtr = get_vtable(t)->template get<CPO>();
-      static_assert(!NoExcept || noexcept(fnPtr(
-          (cpo_t&&) cpo,
-          replace_this<Args>::get((decltype(args) &&) args, objPtr)...)));
-      return fnPtr(
-          (cpo_t&&) cpo,
-          replace_this<Args>::get((decltype(args) &&) args, objPtr)...);
-    }
-  };
-};
-
-template <
-    typename Derived,
-    typename CPO,
-    typename Ret,
-    typename... Args>
-struct _with_type_erased_tag_invoke<
-    Derived,
-    CPO,
-    false,
-    Ret(Args...) noexcept>
-  : _with_type_erased_tag_invoke<Derived, CPO, true, Ret(Args...)> {
-};
 
 template <
     typename Derived,
@@ -409,7 +218,10 @@ class _byval<CPOs...>::type
   }
 
  private:
-  using vtable_holder_t = vtable_holder<_deallocate_cpo, CPOs...>;
+  using vtable_holder_t = std::conditional_t<
+    (sizeof...(CPOs) <= 2),
+    detail::inline_vtable_holder<_deallocate_cpo, CPOs...>,
+    detail::indirect_vtable_holder<_deallocate_cpo, CPOs...>>;
 
   UNIFEX_ALWAYS_INLINE void unsafe_deallocate() noexcept {
     // This leaves the any_unique in an invalid state.
@@ -437,52 +249,7 @@ class _byval<CPOs...>::type
   vtable_holder_t vtable_;
 };
 
-template <typename... CPOs>
-struct _byref {
-  class type;
-};
 
-template <typename... CPOs>
-class _byref<CPOs...>::type
-  : private with_type_erased_tag_invoke<type, CPOs>... {
- public:
-  template (typename Concrete)
-    (requires (!same_as<Concrete const, type const>))
-  /*implicit*/ type(Concrete& impl)
-    : vtable_(vtable_holder_t::template create<Concrete>())
-    , impl_((void*) std::addressof(impl)) {}
-
-  void swap(type& other) noexcept {
-    std::swap(vtable_, other.vtable_);
-    std::swap(impl_, other.impl_);
-  }
-
-  // Two any_ref's compare equal IFF they refer to the same object (shallow).
-  friend bool operator==(type const& left, type const& right) noexcept {
-    return left.impl_ == right.impl_;
-  }
-  friend bool operator!=(type const& left, type const& right) noexcept {
-    return !(left == right);
-  }
-
- private:
-  using vtable_holder_t = vtable_holder<CPOs...>;
-
-  friend void swap(type& left, type& right) noexcept {
-    left.swap(right);
-  }
-
-  friend const vtable_holder_t& get_vtable(const type& self) noexcept {
-    return self.vtable_;
-  }
-
-  friend void* get_object_address(const type& self) noexcept {
-    return self.impl_;
-  }
-
-  void* impl_;
-  vtable_holder_t vtable_;
-};
 
 } // namespace _any_unique
 
@@ -491,12 +258,6 @@ using any_unique = typename _any_unique::_byval<CPOs...>::type;
 
 template <auto&... CPOs>
 using any_unique_t = any_unique<tag_t<CPOs>...>;
-
-template <typename... CPOs>
-using any_ref = typename _any_unique::_byref<CPOs...>::type;
-
-template <auto&... CPOs>
-using any_ref_t = any_ref<tag_t<CPOs>...>;
 
 } // namespace unifex
